feat: Improve splash screen animation
Refactor the delay duration in the initState method of the _SplashViewState class to improve the splash screen animation. The delay is reduced from 5 seconds to 4.5 seconds for a smoother transition. Additionally, a blur animation is added to the splash screen image, starting from an offset of (40, 40) and ending at Offset.zero, with a decelerating curve.